requirements.txt
chromadb==1.4.1
dashscope==1.25.5
pypdf==6.6.2
python_docx==1.2.0
PyYAML==6.0.3
Requests==2.32.5
